--- Paginas/Paginas/css/home.css
@@ -1,189 +1,0 @@
-    body {
-        background-color: #E8AF6B;
-        margin: 0;
-        font-family: "Arial", sans-serif;
-    }
-
-    .container {
-        box-shadow: black;
-        height: 186px;
-        width: 390px;
-        position: absolute;
-        top: 50%;
-        left: 50%;
-        transform: translate(-50%, -50%);
-        text-align: center;
-    }
-
-    .botao-baixo {
-        height: 33%;
-        width: 100%;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-    }
-
-    .botao-alfabeto {
-        width: 100%;
-        height: 33%;
-        margin: 3px;
-        background-color: #D38233;
-        color: white;
-        display: flex;
-        align-items: center;
-        justify-content: space-around;
-        font-family: "Arial", sans-serif;
-        font-weight: 600;
-        font-size: 18px;
-        border-radius: 15px;
-        display: flex;
-        align-items: center;
-        justify-content: space-between; 
-    }
-    
-    .botao-alfabeto button {
-        background-color: #D38233;
-        padding: 10px 20px;
-        border: none;
-        border-radius: 4px;
-        cursor: pointer;
-        transition: background-color 0.3s;
-        display: flex;
-        align-items: center;
-    }
-    
-    .slides {
-        display: flex;
-        flex-direction: column;
-        align-items: center; /* Adicionado para centralizar os slides */
-    }
-    
-    .slide {
-        display: none;
-        text-align: center;
-    }
-    
-    .ativo {
-        display: block;
-    }
-    
-    
-    .botao-dificuldade {
-        width: 100%;
-        color: white;
-        height: 33%;
-        margin: 3px;
-        background-color: #D38233;
-        display: flex;
-        align-items: center;
-        justify-content: space-around;
-        transition: background-color 0.3s;
-        font-family: "Arial", sans-serif;
-        font-weight: 600;
-        font-size: 18px;
-        border-radius: 15px;
-        display: flex;
-        align-items: center;
-        justify-content: space-between; 
-    }
-    
-    .botao-dificuldade button {
-        background-color: #D38233;
-        padding: 10px 20px;
-        border: none;
-        border-radius: 4px;
-        cursor: pointer;
-        display: flex;
-        align-items: center;
-    }
-    
-    .slides {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-    }
-    
-    .slide {
-        display: none;
-        text-align: center;
-    }
-    
-    .ativo {
-        display: block;
-    }
-    
- 
-    
-
-    .botao_ciculos button {
-        width: 116px;
-        height: 60px;
-        margin: 3px;
-        background-color: #D38233;
-        color: white;
-        padding: 10px 20px;
-        border: none;
-        border-radius: 4px;
-        cursor: pointer;
-        transition: background-color 0.3s;
-        border-radius: 15px;
-    }
-
-    .botao-inicio button {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        flex-direction: column;
-        text-align: center;
-        width: 146px;
-        height: 60px;
-        margin: 3px;
-        background-color: #D38233;
-        padding: 10px 20px;
-        border: none;
-        border-radius: 4px;
-        cursor: pointer;
-        transition: background-color 0.3s;
-        border-radius: 15px;
-    }
-
-    .botao-inicio svg {
-        width: 65px;
-        height: 65px;
-        margin-bottom: 5px;
-        fill: white;
-    }
-
-    .botao-dicionario button {
-        width: 116px;
-        height: 60px;
-        margin: 3px;
-        background-color: #D38233;
-        color: white;
-        padding: 10px 20px;
-        border: none;
-        border-radius: 4px;
-        cursor: pointer;
-        transition: background-color 0.3s;
-        border-radius: 15px;
-    }
-
-    .botao-anterior,
-    .botao-proximo {
-        background-color: #D38233;
-        color: #fff;
-        padding: 10px 15px;
-        border: none;
-        cursor: pointer;
-        border-radius: 15px;
-        transition: background-color 0.3s ease;
-    }
-    
-    .botao-anterior svg,
-    .botao-proximo svg {
-        width: 25px;
-        height: 25px;
-        fill: white;
-        margin-right: 5px;
-        
-    }
