--- Paginas/Paginas/hml/desafio.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-BR">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="../css/desafio.css">
-  
-    <title>Tela Enki</title>
-</head>
-<body>
-    <div class="circle"></div>
-    <div class="circle"></div>
-    <div class="circle"></div>
-    <div class="container"></div>
-    <input type="text" id="inputVogal" placeholder="Digite a vogal correspondente em português">
-    <script src="../js/vetor.js"></script> <!-- Importar vetor.js antes de desafio.js -->
-    <script src="../js/desafio.js"></script>
-</body>
-</html>
