+++ Paginas/css/desafio.css
@@ -31,25 +31,4 @@ input {
   font-size: 16px;
 }
 
-.circle {
-  position: absolute;
-  width: 25px;
-  height: 25px;
-  background: #ffffff;
-  border-radius: 50%;
-  top: calc(100% - 580px); /* Posição vertical centralizada */
-}
-
-.circle:nth-child(1) {
-  left: calc(50% - 55px); /* Posição horizontal ajustada */
-}
-
-.circle:nth-child(2) {
-  left: calc(50% - 15px); /* Posição horizontal ajustada */
-}
-
-.circle:nth-child(3) {
-  left: calc(50% + 25px); /* Posição horizontal ajustada */
-}
-
