+++ Paginas/css/front.css
@@ -10,16 +10,9 @@ body {
     transform: translate(-50%, -50%);
 }
 
-.botao-baixo {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  
-}
-
 .botao-alfabeto button {
     width: 390px; 
-    height: 60px;
+    height: 70px;
     margin: 3px;
     background-color: #D38233;
     color: white;
@@ -36,7 +29,7 @@ body {
 
 .botao-dificuldade button{
     width: 390px; 
-    height: 60px;
+    height: 70px;
     margin: 3px;
     background-color: #D38233;
     color: white;
