+++ Paginas/hml/desafio.html
@@ -8,12 +8,8 @@
     <title>Tela Enki</title>
 </head>
 <body>
-    <div class="circle"></div>
-    <div class="circle"></div>
-    <div class="circle"></div>
     <div class="container"></div>
-    <input type="text" id="inputVogal" placeholder="Digite a vogal correspondente em português">
-    <script src="../js/vetor.js"></script> <!-- Importar vetor.js antes de desafio.js -->
+    <input type="text" id="inputVogal" placeholder="Digite a resposta">
     <script src="../js/desafio.js"></script>
 </body>
 </html>
