+++ Paginas/hml/desafio2.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="pt-BR">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="../css/desafio.css">
+  
+    <title>Tela Enki</title>
+</head>
+<body>
+    <div class="container"></div>
+    <input type="text" id="inputVogal" placeholder="Digite a resposta">
+    <script src="../js/desafio2.js"></script>
+</body>
+</html>
