+++ Paginas/hml/front.html
@@ -9,26 +9,22 @@
 <body>
     <div class="container">
         <div class="botao-alfabeto"> 
-            <button type="button">ひらがな<br>Hiragana</button>
+            <button type="button" onclick="redirecionarParaHiragana()">ひらがな<br>Hiragana</button>
         </div>
         <div class="botao-dificuldade">
-            <button type="button">Basico</button>
+            <button type="button" onclick="redirecionarParaKatakana()">カタカナ<br>Katakana</button>
         </div>
-        <div class="botao-baixo">
-            <div class="botao_ciculos">
-                <button type="button">000</button>
-            </div>
-            <div class="botao-inicio">
-                <button type="button">
-                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width="40" height="40">
-                        <path d="M16.3944 11.9998L10 7.73686V16.2628L16.3944 11.9998ZM19.376 12.4158L8.77735 19.4816C8.54759 19.6348 8.23715 19.5727 8.08397 19.3429C8.02922 19.2608 8 19.1643 8 19.0656V4.93408C8 4.65794 8.22386 4.43408 8.5 4.43408C8.59871 4.43408 8.69522 4.4633 8.77735 4.51806L19.376 11.5838C19.6057 11.737 19.6678 12.0474 19.5146 12.2772C19.478 12.3321 19.4309 12.3792 19.376 12.4158Z" fill="white"/>
-                    </svg>
-                     
-            </div>
-            <div class="botao-dicionario">
-                <button type="button">dicionario</button>
-            </div>
     </div>
+
+    <script>
+        function redirecionarParaHiragana() {
+            // Redireciona para a página desejada
+            window.location.href = "desafio.html";
+        }
+        function redirecionarParaKatakana() {
+            // Redireciona para a página desejada
+            window.location.href = "desafio2.html";
+        }
+    </script>
 </body>
 </html>
-
