--- Paginas/js/vetor.js
@@ -1,153 +1,0 @@
-var portugues = [
-    'a', 'i', 'u', 'e', 'o',
-    'ka', 'ki', 'ku', 'ke', 'ko',
-    'sa', 'shi', 'su', 'se', 'so',
-    'ta', 'chi', 'tsu', 'te', 'to',
-    'na', 'ni', 'nu', 'ne', 'no',
-    'ha', 'hi', 'fu', 'he', 'ho',
-    'ma', 'mi', 'mu', 'me', 'mo',
-    'ya', 'yu', 'yo',
-    'ra', 'ri', 'ru', 're', 'ro',
-    'wa', 'wo', 'n',
-    'ga', 'gi', 'gu', 'ge', 'go',
-    'za', 'ji', 'zu', 'ze', 'zo',
-    'da', 'ji', 'zu', 'de', 'do',
-    'ba', 'bi', 'bu', 'be', 'bo',
-    'pa', 'pi', 'pu', 'pe', 'po',
-    'kya', 'kyu', 'kyo',
-    'sha', 'shu', 'sho',
-    'cha', 'chu', 'cho',
-    'nya', 'nyu', 'nyo',
-    'hya', 'hyu', 'hyo',
-    'mya', 'myu', 'myo',
-    'rya', 'ryu', 'ryo',
-    'gya', 'gyu', 'gyo',
-    'ja', 'ju', 'jo',
-    'ja', 'ju', 'jo',
-    'bya', 'byu', 'byo',
-    'pya', 'pyu', 'pyo'
-];
-
-var hiragana_baisoc = ['あ', 'い', 'う', 'え', 'お',
-            'か', 'き', 'く', 'け', 'こ',
-            'さ', 'し', 'す', 'せ', 'そ',
-            'た', 'ち', 'つ', 'て', 'と',
-            'な', 'に', 'ぬ', 'ね', 'の',
-            'は', 'ひ', 'ふ', 'へ', 'ほ',
-            'ま', 'み', 'む', 'め', 'も',
-            'や', 'ゆ', 'よ',
-            'ら', 'り', 'る', 'れ', 'ろ',
-            'わ', 'を', 'ん'];
-
-
-
-var hiragana_intermediario = [
-            'あ', 'い', 'う', 'え', 'お',
-            'か', 'き', 'く', 'け', 'こ',
-            'さ', 'し', 'す', 'せ', 'そ',
-            'た', 'ち', 'つ', 'て', 'と',
-            'な', 'に', 'ぬ', 'ね', 'の',
-            'は', 'ひ', 'ふ', 'へ', 'ほ',
-            'ま', 'み', 'む', 'め', 'も',
-            'や', 'ゆ', 'よ',
-            'ら', 'り', 'る', 'れ', 'ろ',
-            'わ', 'を', 'ん',
-            'が', 'ぎ', 'ぐ', 'げ', 'ご',
-            'ざ', 'じ', 'ず', 'ぜ', 'ぞ',
-            'だ', 'ぢ', 'づ', 'で', 'ど',
-            'ば', 'び', 'ぶ', 'べ', 'ぼ',
-            'ぱ', 'ぴ', 'ぷ', 'ぺ', 'ぽ'];
-
-
-var hiragana_avancado = [
-    'あ', 'い', 'う', 'え', 'お',
-    'か', 'き', 'く', 'け', 'こ',
-    'さ', 'し', 'す', 'せ', 'そ',
-    'た', 'ち', 'つ', 'て', 'と',
-    'な', 'に', 'ぬ', 'ね', 'の',
-    'は', 'ひ', 'ふ', 'へ', 'ほ',
-    'ま', 'み', 'む', 'め', 'も',
-    'や','ゆ', 'よ',
-    'ら', 'り', 'る', 'れ', 'ろ',
-    'わ','を', 'ん',
-    'が', 'ぎ', 'ぐ', 'げ', 'ご',
-    'ざ', 'じ', 'ず', 'ぜ', 'ぞ',
-    'だ', 'ぢ', 'づ', 'で', 'ど',
-    'ば', 'び', 'ぶ', 'べ', 'ぼ',
-    'ぱ', 'ぴ', 'ぷ', 'ぺ', 'ぽ',
-    'きゃ', 'きゅ', 'きょ',
-    'しゃ', 'しゅ', 'しょ',
-    'ちゃ', 'ちゅ', 'ちょ',
-    'にゃ', 'にゅ', 'にょ',
-    'ひゃ', 'ひゅ', 'ひょ',
-    'みゃ', 'みゅ', 'みょ',
-    'りゃ', 'りゅ', 'りょ',
-    'ぎゃ', 'ぎゅ', 'ぎょ',
-    'じゃ', 'じゅ', 'じょ',
-    'ぢゃ', 'ぢゅ', 'ぢょ',
-    'びゃ', 'びゅ', 'びょ',
-    'ぴゃ', 'ぴゅ', 'ぴょ'
-];
-
-var katakana_basico = [
-                    'ア', 'イ', 'ウ', 'エ', 'オ',
-                   'カ', 'キ', 'ク', 'ケ', 'コ',
-                   'サ', 'シ', 'ス', 'セ', 'ソ',
-                   'タ', 'チ', 'ツ', 'テ', 'ト',
-                   'ナ', 'ニ', 'ヌ', 'ネ', 'ノ',
-                   'ハ', 'ヒ', 'フ', 'ヘ', 'ホ',
-                   'マ', 'ミ', 'ム', 'メ', 'モ',
-                   'ヤ', 'ユ', 'ヨ',
-                   'ラ', 'リ', 'ル', 'レ', 'ロ',
-                   'ワ', 'ヲ', 'ン'];
-
-var katakana_intermediario = [
-    'ア', 'イ', 'ウ', 'エ', 'オ',
-    'カ', 'キ', 'ク', 'ケ', 'コ',
-    'サ', 'シ', 'ス', 'セ', 'ソ',
-    'タ', 'チ', 'ツ', 'テ', 'ト',
-    'ナ', 'ニ', 'ヌ', 'ネ', 'ノ',
-    'ハ', 'ヒ', 'フ', 'ヘ', 'ホ',
-    'マ', 'ミ', 'ム', 'メ', 'モ',
-    'ヤ', 'ユ', 'ヨ',
-    'ラ', 'リ', 'ル', 'レ', 'ロ',
-    'ワ', 'ヲ', 'ン',
-    'ガ', 'ギ', 'グ', 'ゲ', 'ゴ',
-    'ザ', 'ジ', 'ズ', 'ゼ', 'ゾ',
-    'ダ', 'ヂ', 'ヅ', 'デ', 'ド',
-    'バ', 'ビ', 'ブ', 'ベ', 'ボ',
-    'パ', 'ピ', 'プ', 'ペ', 'ポ'
-];
-
-var katakana_avancado = [
-    'ア', 'イ', 'ウ', 'エ', 'オ',
-    'カ', 'キ', 'ク', 'ケ', 'コ',
-    'サ', 'シ', 'ス', 'セ', 'ソ',
-    'タ', 'チ', 'ツ', 'テ', 'ト',
-    'ナ', 'ニ', 'ヌ', 'ネ', 'ノ',
-    'ハ', 'ヒ', 'フ', 'ヘ', 'ホ',
-    'マ', 'ミ', 'ム', 'メ', 'モ',
-    'ヤ', 'ユ', 'ヨ',
-    'ラ', 'リ', 'ル', 'レ', 'ロ',
-    'ワ', 'ヲ', 'ン',
-    'ガ', 'ギ', 'グ', 'ゲ', 'ゴ',
-    'ザ', 'ジ', 'ズ', 'ゼ', 'ゾ',
-    'ダ', 'ヂ', 'ヅ', 'デ', 'ド',
-    'バ', 'ビ', 'ブ', 'ベ', 'ボ',
-    'パ', 'ピ', 'プ', 'ペ', 'ポ',
-    'キャ', 'キュ', 'キョ',
-    'シャ', 'シュ', 'ショ',
-    'チャ', 'チュ', 'チョ',
-    'ニャ', 'ニュ', 'ニョ',
-    'ヒャ', 'ヒュ', 'ヒョ',
-    'ミャ', 'ミュ', 'ミョ',
-    'リャ', 'リュ', 'リョ',
-    'ギャ', 'ギュ', 'ギョ',
-    'ジャ', 'ジュ', 'ジョ',
-    'ヂャ', 'ヂュ', 'ヂョ',
-    'ビャ', 'ビュ', 'ビョ',
-    'ピャ', 'ピュ', 'ピョ'
-];
-
-
-
