--- teste.js
@@ -1,4 +1,0 @@
-// outroArquivo.js
-
-const getCarcterHB = require('../Anki/index');
-console.log(getCarcterHB())
